Read time series data from SWMM's .out file

Reads the binary output ('.out') generated by the stormwater management model 'SWMM' and creates a list of xts-objects.

read_out(
  file = "",
  iType = NULL,
  object_name = NULL,
  vIndex = NULL,
  firstPeriod = NULL,
  lastPeriod = NULL,
  multiColumn = FALSE,
  byObject = TRUE,
  method = 1L
)

Arguments

file

The file to be read.

iType

Sets the result type: 0 for Subcatchments, 1 for nodes, 2 for links, 3 for system variables. Leave empty for retrieving elements available.

object_name

Sets the objects of which time series data is returned. Leave empty for retrieving elements available.

vIndex

Sets the variables to be read (s. Details). Leave empty for retrieving elements available.

firstPeriod

integer number of first period to be returned

lastPeriod

integer number of last period to be returned

multiColumn

if TRUE (the default is FALSE), the result is a list of xts objects with each object containing all variables instead of a list of lists of xts objects.

byObject

if TRUE, each top level list element represents an object, otherwise a variable

method

Integer number determining the method of reading the binary file. The default is method = 1. If multiColumn = TRUE and byObject = TRUE you may try method = 2 and see if performance is enhanced (intended effect) compared to method = 1. The results should be the same!

Value

A list of a list of xts-objects (if multiColumn = TRUE) or a list of xts-objects (if multiColumn = FALSE).

Details

vIndex depends on the result type. Choices are...

for each subcatchment variable:

0 for rainfall rate (in/hr or mm/hr),

1 for snow depth (inches or millimeters),

2 for evaporation loss (in/day or mm/day),

3 for infiltration loss (in/hr or mm/hr),

4 for runoff flow (flow units),

5 for groundwater flow into the drainage network (flow units),

6 for groundwater elevation (ft or m),

7 for soil moisture in the unsaturated groundwater zone (volume fraction),

7 + N for washoff concentration of each pollutant (mass/liter).

for each node variable:

0 for water depth (ft or m above the node invert elevation),

1 for hydraulic head (ft or m, absolute elevation per vertical datum),

2 for stored water volume (including ponded water, ft3 or m3),

3 for lateral inflow (runoff + all other external inflows, in flow units),

4 for total inflow (lateral inflow + upstream inflows, in flow units),

5 for surface flooding (excess overflow when the node is at full depth, in flow units),

5 + N for concentration of each pollutant after any treatment (mass/liter),

for each link variable:

0 for flow rate (flow units),

1 for average water depth (ft or m),

2 for flow velocity (ft/s or m/s),

3 for volume of water (ft3 or m3),

4 for capacity (fraction of full area filled by flow for conduits; control setting for pumps and regulators),

4 + N for concentration of each pollutant (mass/liter),

for each system-wide variable:

0 for air temperature (deg. F or deg. C),

1 for total rainfall (in/hr or mm/hr),

2 for total snow depth (inches or millimeters),

3 for average losses (in/hr or mm/hr),

4 for total runoff (flow units),

5 for total dry weather inflow (flow units),

6 for total groundwater inflow (flow units),

7 for total RDII inflow (flow units),

8 for total external inflow (flow units),

9 for total direct inflow (flow units),

10 for total external flooding (flow units),

11 for total outflow from outfalls (flow units),

12 for total nodal storage volume ( ft3 or m3),

13 for potential evaporation (in/day or mm/day),

14 for actual evaporation (in/day or mm/day).

See also

Examples

 
if (FALSE) {
xts_list_of_results <- read_out("model.out")
}